app/requirements.txt
slack_bolt==1.19.0
psycopg2-binary==2.9.9
requests==2.32.3
